// src/types/bias.ts
export interface CognitiveBias {
  Bias: string;
  Descrizione: string;
  "Tecniche per gestirlo": string;
  "Contestualizzazione nel Reale": string;
  Vantaggi: string;
  Svantaggi: string;
  "Attività Esperienziale di Aula": string;
  Fonti: string;
  "TAG 1": string;
  TAG2: string;
  "TAG LUSTER": string;
}

export interface SearchFilters {
  nameQuery: string;
  descriptionQuery: string;
  selectedTags: string[];
}